Add hacky assembly resolution and get NuGet package working.
When the type provider runs in fsc.exe or in the compiler service for
intellisense, we have to go poking around parent folders to see if we
are installed as a NuGet package and pull in our dependencies from the
packages folder. This is not pretty but seems to work.