Make default constraint names table-prefixed for all backends.
Previously I had thought that SQL Server was weird having constraints be global
to the database. Indeed, Postgres technically has constraint names local to the
table. But, it seems that constraints that require an associated index like
UNIQUE and PRIMARY KEY constraints will get that index named matching the
constraint, and the index name is scoped to the database, so this leads to
collisions.

This will be a cleaner approach overall anyway, since it leads to less surprises
on TSQL when you use a custom constraint name and the translator hackily adds a
table prefix to it.

We still need to check constraint uniqueness in ModelOps so we can error when it
is violated.